refactor(settings): split setting registry by section
